requirements.txt

google-cloud-error-reporting==0.34.0
google-cloud-pubsub==1.7.0
google-cloud-storage==1.30.0
google-cloud-translate==2.0.2
google-cloud-vision==1.0.0
google-cloud-firestore==1.8.1